Docs: slider 폴더 생성
